Implement push notifications (#2)
* WIP: Push notifications

* WIP: Integrate with Firebase

* Store notification subscriptions

* Update filters when subscription already exists

* Send push notifications

* Generate notification text

* Update click action

* Refactor notification

* Improve errors

* Get found apartments

* Improve routes

* Update packages

* Refactor

* Add default sorting for apartment entity

* Get found apartments

* Refactor

* Remove unused code

* Refactor

* Fix types

* Improve pagination

* Improve notification subscription response

* Filter found apartments by date

* Refactor page size

* Handle unsubscription

* Refactor notification subscription service

* Add tests

* Add tests

* Add some e2e tests

* Add some tests

* Add tests

* Add tests for subscriptions

* Add e2e tests

* Fix tests

* Fix tests

* Fix tests

* Add new environment variable

* Format

* Format

* Add more tests

* Add test for subscription

* Add more tests

+++ src/modules/tasks/tasks.service.ts
@@ -1,15 +1,18 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
 import { Transactional } from 'typeorm-transactional-cls-hooked';
-import { RECEIVED_APARTMENTS_SIZE } from 'modules/apartment/apartment.constants';
 import { ApartmentService } from 'modules/apartment/apartment.service';
 import { FilterDto } from 'modules/filter/dto/filter.dto';
 import { filters } from 'modules/filter/filter.constants';
 import { Filter } from 'modules/filter/filter.interface';
 import { FilterService } from 'modules/filter/filter.service';
 import { MailService } from 'modules/mail/mail.service';
-import { defaultPaginationParams } from 'modules/pagination/pagination.constants';
+import {
+  defaultPaginationParams,
+  DEFAULT_LIMIT_PER_PAGE,
+} from 'modules/pagination/pagination.constants';
 import { getSkip } from 'modules/pagination/pagination.utils';
+import { SubscriptionService } from 'modules/subscription/subscription.service';
 import { FILTER_DEACTIVATION_TOKEN_EXPIRATION_HOURS } from 'modules/token/token.constants';
 import { TokenType } from 'modules/token/token.enums';
 import { TokenService } from 'modules/token/token.service';
@@ -25,6 +28,7 @@ export class TasksService {
     private readonly apartmentService: ApartmentService,
     private readonly filterService: FilterService,
     private readonly mailService: MailService,
+    private readonly subscriptionService: SubscriptionService,
     private readonly tokenService: TokenService,
     private readonly userService: UserService,
   ) {}
@@ -106,11 +110,15 @@ export class TasksService {
 
   @Transactional()
   private async sendNewApartmentsByFilter(filter: Filter): Promise<void> {
-    await this.tokenService.deleteTokenByFilterId(filter.id);
+    const userEmail = await this.userService.getUserEmail(filter.userId);
+
+    if (userEmail) {
+      await this.tokenService.deleteTokenByFilterId(filter.id);
+    }
 
     const apartmentList = await this.apartmentService.getApartmentListFromDatabaseByFilter(
       filter,
-      RECEIVED_APARTMENTS_SIZE,
+      DEFAULT_LIMIT_PER_PAGE,
     );
     if (apartmentList.data.length === 0) {
       return;
@@ -121,21 +129,28 @@ export class TasksService {
         firstApartment.price - secondApartment.price,
     );
 
-    const filterDeactivationUrl = await this.filterService.createTokenAndDeactivationUrl(
-      {
-        filterId: filter.id,
-        userId: filter.userId,
-        type: TokenType.DEACTIVATION,
-      },
-      FILTER_DEACTIVATION_TOKEN_EXPIRATION_HOURS,
-    );
-    const userEmail = await this.userService.getUserEmail(filter.userId);
-    await this.mailService.sendMailWithNewApartments(
-      userEmail,
-      newApartments,
-      filter as FilterDto,
-      filterDeactivationUrl,
-    );
+    if (userEmail) {
+      const filterDeactivationUrl = await this.filterService.createTokenAndDeactivationUrl(
+        {
+          filterId: filter.id,
+          userId: filter.userId,
+          type: TokenType.DEACTIVATION,
+        },
+        FILTER_DEACTIVATION_TOKEN_EXPIRATION_HOURS,
+      );
+      await this.mailService.sendMailWithNewApartments(
+        userEmail,
+        newApartments,
+        filter as FilterDto,
+        filterDeactivationUrl,
+      );
+    } else {
+      const isNotificationSent = await this.subscriptionService.sendNotification(
+        filter,
+        newApartments.length,
+      );
+      if (!isNotificationSent) return;
+    }
 
     await this.userService.insertReceivedApartments(
       filter.userId,
